+++ NGP/NGP_Client/Bird.cpp
@@ -117,5 +117,5 @@ void Bird::Decode(ObjectData object_data)
 {
 	Enemy::SetLocaition((int)object_data.pos.x, (int)object_data.pos.y);
 	Enemy::SetHp((int)object_data.hp);
-	Enemy::SetAct((int)object_data.state);
+	Enemy::SetAct(object_data.state);
 }
